Changes to be committed: deleted:    README 	modified:   README.md 	deleted:    src/.gitignore 	modified:   src/mysql.cpp 	modified:   src/nanny.cpp 	modified:   src/protos.h 	modified:   src/server.cpp
Readme updated to me more inclusive for a droplet setup.

server.cpp modified so that it has a default configuration that will
immediately work if droplet instructions are followed.

mysql, nanny, and protos .cpp files were modified to comment out family
clanning. This isn't an ideal fix, but it does stop it from crashing the
server (so far).

.gitignore file removed from src directory as the directory (copy) I use
to commit changes isn't also used for compiling and running the server.